Management - refactoring, move api layers into modules
