+++ management-service/src/modules/customers/api/controllers/command/index.ts
