+++ management-service/src/modules/customers/api/controllers/query/index.ts
