+++ management-service/src/modules/enterprise/api/schemas/CreateEnterprise.schema.ts
@@ -4,8 +4,7 @@ import { TextValidator } from 'shared/core';
 import { contactPersonSchema } from 'shared/domain/schemas';
 
 import { CreateEnterpriseDto } from 'modules/enterprise/application/command/createEnterprise';
-
-import { countryCodes } from '../../../resources/countryCodes';
+import { countryCodes } from 'resources/countryCodes';
 
 export const createEnterpriseSchema = yup.object().shape<CreateEnterpriseDto>({
   enterpriseName: yup.string().required().min(1).max(999),
