+++ management-service/src/modules/facilities/api/controllers/command/RemoveEmployee.controller.ts
