+++ management-service/src/modules/facilities/api/controllers/command/RemoveFacility.controller.ts
