+++ management-service/src/modules/facilities/api/controllers/command/RemoveOffer.controller.ts
