+++ management-service/src/modules/facilities/api/controllers/query/GetFacilityById.controller.ts
