+++ management-service/src/modules/facilities/api/controllers/query/GetFacilityBySlug.controller.ts
